// vhome_mis/src/app/core/api-client-v2/model/employeeData.ts
/**
 * APIFlask
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface EmployeeData { 
    ID: number;
    company_email: string;
    date_of_employment: string;
    deleted: boolean;
    localization: string;
    name: string;
}